+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [Unreleased]
 
+### Fixes
+
+- [CM-122] Updated CSAT rating scale to 1-10 from 1-3.
+
 ## [3.1.0] - 2020-01-28
 
 ### Enhancements